Feature/convert rewards to dyna

+++ contracts/dynaBNB/interfaces/IPancakeRouter.sol
@@ -48,5 +48,8 @@ interface IPancakeRouter {
         uint256 deadline
     ) external payable returns (uint256[] memory amounts);
 
-    function getAmountsOut(uint amountIn, address[] calldata path) external view returns (uint256[] memory amounts);
+    function getAmountsOut(uint256 amountIn, address[] calldata path)
+        external
+        view
+        returns (uint256[] memory amounts);
 }
